Refactor project into packages
